Move flag computation to interner
